agent_using_mcp/requirements.txt
langchain
langgraph
langchain-google-genai
python-dotenv
requests